refactor: move sender account state to new file
